Initial implementation of the Text or Binary algorithm
This is a very simply written implementation of the HTML5 algorithm
that sniffs content when the server says it is 'text/plain'. It
detects if the content seems to be binary using a simple test table,
and avoids privilege escalation from text/plain to types known to be
scriptable. Tests included.