Use GIO to further sniff unknown content
The HTML5 spec allows user agents to further sniff the content if the
normal 'unknown' algorithm isn't able to find a type, before going
with 'application/octet-stream'. This change uses the already existing
sniff_gio code to perform that sniffing, and also fixes the default
behavior of sniffing to be return the type given by the server.