soup-session-async: move cache handling to run_queue
We were queueing an idle function queue_message before, but this can
fail in the following way: if run_queue is somehow called before the
scheduled idle runs, it will find the message you intended to reply to
from the cache in the queue, and will request it to the server anyway.

Now we reply directly when iterating the queue, and set the IOStatus
to RUNNING so that we don't reply more than once.